Wired in calendar editing

+++ renderer/src/components/EditCalendar.tsx
@@ -16,7 +16,7 @@ along with Contact Schedular.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import * as React from 'react';
-import { ICalendar } from '../common/types';
+import { ICalendar, CalendarSource } from '../common/types';
 
 export interface IStateProps {
   calendar: ICalendar;
@@ -26,6 +26,7 @@ export interface IStateProps {
 export interface IDispatchProps {
   saveCalendar: (calendar: ICalendar) => void;
   deleteCalendar: (calendar: ICalendar) => void;
+  closeCalendar: () => void;
 }
 
 export type IProps = IStateProps & IDispatchProps;
@@ -45,13 +46,26 @@ export class EditCalendar extends React.Component<IProps, IState> {
   public render() {
     const entry = this.state.unsavedEntry;
     let deleteButton: JSX.Element | undefined = undefined;
+    let saveButton: JSX.Element | undefined = undefined;
+    let contents: JSX.Element | undefined = undefined;
     if (!this.props.isAdd) {
       deleteButton = (
         <button type="button" className="btn btn-danger btn-lg" onClick={this.onDelete}>Delete</button>
       );
     }
-    return (
-      <div className="edit-calendar-container">
+    if (this.props.isAdd && !this.state.unsavedEntry.source) {
+      contents = (
+        <div className="edit-calendar-data-source">
+          <button type="button" className="btn btn-primary btn-lg" onClick={() => this.onSourceEdit('office365')}>
+            Office 365
+          </button>
+          <button type="button" className="btn btn-primary btn-lg" onClick={() => this.onSourceEdit('google')}>
+            Google
+          </button>
+        </div>
+      );
+    } else {
+      contents = (
         <div className="edit-calendar-data">
           <div>
             <span>Name:</span>
@@ -60,12 +74,35 @@ export class EditCalendar extends React.Component<IProps, IState> {
             <input type="text" value={entry.displayName} onChange={this.onNameEdit} />
           </div>
         </div>
+      );
+      saveButton = (
+        <button type="button" className="btn btn-primary btn-lg" onClick={this.onSave}>Save</button>
+      );
+    }
+    return (
+      <div className="edit-calendar-container">
+        <h3 className="edit-contact-header">{this.props.isAdd ? 'New Calendar' : 'Edit Calendar'}</h3>
+        {contents}
         <div className="edit-calendar-buttons">
           {deleteButton}
-          <button type="button" className="btn btn-primary btn-lg" onClick={this.onSave}>Save</button>
+          {saveButton}
+          <button type="button" className="btn btn-primary btn-lg" onClick={this.onCancel}>Cancel</button>
         </div>
       </div>
     );
+  }
+
+  private onSourceEdit = (source: CalendarSource) => {
+    this.setState((previousState) => {
+      const newState: IState = {
+        ...previousState,
+        unsavedEntry: {
+          ...previousState.unsavedEntry,
+          source
+        }
+      };
+      return newState;
+    });
   }
 
   private onNameEdit = (event: React.FormEvent<HTMLInputElement>) => {
@@ -83,10 +120,17 @@ export class EditCalendar extends React.Component<IProps, IState> {
   }
 
   private onDelete = () => {
-    this.props.deleteCalendar(this.props.calendar);
+    const msg = `Are you sure you would like to delete ${this.props.calendar.displayName} from your contacts list?`;
+    if (window.confirm(msg)) {
+      this.props.deleteCalendar(this.props.calendar);
+    }
   }
 
   private onSave = () => {
     this.props.saveCalendar(this.state.unsavedEntry);
+  }
+
+  private onCancel = () => {
+    this.props.closeCalendar();
   }
 }
